Payment Evol. 4 Totals
Total of the order is now displayed in both menu and payment tabs. This
will update correctly when adding and item, taking an item away and
clearing the order.

+++ src/consumable/Consumable.java
@@ -28,7 +28,7 @@ public class Consumable implements Comparable<Consumable>, Serializable {
   private String name;
 
   /** The price of the dish. */
-  private float price;
+  private double price;
 
   /** Calories of the dish. */
   private int calories;
@@ -132,7 +132,7 @@ public class Consumable implements Comparable<Consumable>, Serializable {
    *
    * @return the price
    */
-  public float getPrice() {
+  public double getPrice() {
     return this.price;
   }
